pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["config", "config.settings", "accounts", "accounts.tests", "library", "library.services", "library.api", "library.tests", "backoffice", "backoffice.tests", "reader", "reader.tests"]

[project]
name = "who-is-rat"
version = "2.0.0"
requires-python = ">=3.12"
dependencies = [
    "Django~=6.0",
    "psycopg[binary]~=3.3",
    "celery~=5.6",
    "redis~=7.3",
    "Pillow~=12.1",
    "argon2-cffi~=25.1",
    "django-redis~=6.0",
    "gunicorn~=25.1",
    "djangorestframework~=3.16",
    "drf-spectacular~=0.28",
    "blind-watermark~=0.4.4",
    "pytesseract~=0.3.13",
    "requests~=2.32",
    "pypinyin>=0.53",
    "opencv-python-headless>=4.9",
    "numpy>=1.26",
]

[project.optional-dependencies]
dev = [
    "pytest~=9.0",
    "pytest-django~=4.12",
    "factory-boy~=3.3",
    "coverage~=7.7",
    "ruff~=0.11",
    "pre-commit~=4.2",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "C4", "SIM"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["test_*.py"]
addopts = "--tb=short -q"

[tool.coverage.run]
source = ["accounts", "library", "backoffice", "reader"]
omit = ["*/tests/*", "*/migrations/*"]
